feat js statJSP

+++ src/main/resources/init.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS tas_user( -- user 테이블
 -- 접속이력 테이블
 CREATE TABLE IF NOT EXISTS tas_logs
 (
-	log_id BIGINT(20) AUTO_INCREMENT NOT NULL PRIMARY KEY COMMENT '로그 아이디',
+	log_id INT(30) AUTO_INCREMENT NOT NULL PRIMARY KEY COMMENT '로그 아이디',
 	ip VARCHAR(50) COMMENT '아이피',
 	url VARCHAR(200) COMMENT '요청 url',
 	http_method VARCHAR(10) COMMENT 'http method',
